+++ src/compute-engine/numerics/numeric-bignum.ts
@@ -11,7 +11,10 @@ export function lcm(a: BigNum, b: BigNum): BigNum {
   return a.mul(b).div(gcd(a, b));
 }
 
-export function factorial2(ce: IBigNum, n: BigNum): BigNum {
+export function* factorial2(
+  ce: IBigNum,
+  n: BigNum
+): Generator<BigNum, BigNum> {
   if (!n.isInteger() || n.isNegative()) return ce._BIGNUM_NAN;
   if (n.lessThan(1)) return ce._BIGNUM_ONE;
 
@@ -19,6 +22,7 @@ export function factorial2(ce: IBigNum, n: BigNum): BigNum {
   while (n.greaterThan(2)) {
     n = n.minus(2);
     result = result.mul(n);
+    yield result;
   }
 
   return result;